Api-roll 21: 2015-08-04
Fixed a bug in the package configuration handling which caused all
pacakges not in config.yaml to be ignored. This resulted in a number
of new APIs and the re-appearence of APIs which was previously
ignored. The real new APIs have been added and the previously ignored
APIs are ignored again.

R=kustermann@google.com, wibling@google.com

Review URL: https://codereview.chromium.org//1268013003 .